+++ CHANGELOG.md
@@ -18,6 +18,14 @@ Types of changes:
 
 ## [Unreleased]
 
+### Added
+
+* Improved documentation for search functions (now in `konsepy.rxsearch`) in README.
+
+### Changed
+
+* Created `konsepy.rxsearch` module to house regex search functions, and create separate tests for them (no longer in `regex`)
+
 ## [0.2.3] - 2026-01-29
 
 ### Added